revert commit 61bb8b2
(61bb8b2)  Ian Remmler, 7 months ago: Fix x coordinate calculation in buttonpress.

+++ dwm.c
@@ -628,7 +628,7 @@ buttonpress(XEvent *e)
 			arg.ui = 1 << i;
 		} else if (ev->x < x + blw)
 			click = ClkLtSymbol;
-		else if (ev->x > (x = selmon->ww - TEXTW(stext) + lrpad)) {
+		else if (ev->x > (x = selmon->ww - (int)TEXTW(stext) + lrpad)) {
 			click = ClkStatusText;
 
 			char *text = rawstext;
